Share the dist asset cache between the two example servers
The streaming example shipped its own copy of the bounded `dist/` cache the
commerce example already had, and said so in a comment. Two of ten examples
hand-rolling the same traversal is the point at which it should be shared
rather than copied a third time.

`examples/shared/dist-assets` now owns the parts that are genuinely identical:
the iterative directory walk, the MIME lookup, and esbuild content-hash
detection. It deliberately takes no web-framework dependency, because the two
callers disagree about cache policy and do not share HTTP types -- commerce
serves unhashed files `public, max-age=86400` while streaming serves them
`no-cache`, and both policies are preserved exactly.

What stays in each example is the part that is actually its own: commerce keeps
`serve_asset` and its separate CSS map, streaming keeps `asset_response` and
`insert_generated_css`.

Net -209 lines across the two examples against 181 in the shared crate, and the
duplicated `is_content_hashed` tests collapse into one set that covers both
examples' filenames.
