+++ examples/app/streaming/server/Cargo.toml
@@ -13,6 +13,7 @@ serde_json = { workspace = true }
 microsoft-webui = { path = "../../../../crates/webui" }
 microsoft-webui-handler = { path = "../../../../crates/webui-handler" }
 microsoft-webui-tokens = { path = "../../../../crates/webui-tokens" }
+webui-example-dist-assets = { path = "../../../shared/dist-assets" }
 actix-web = { workspace = true }
 anyhow = { workspace = true }
 bytes = { workspace = true }
@@ -20,7 +21,6 @@ tokio = { workspace = true }
 tokio-stream = { workspace = true }
 futures-util = { workspace = true }
 clap = { workspace = true }
-mime_guess = { workspace = true }
 
 [lints]
 workspace = true
